// apps/template/src/api/generated/types.gen.ts
export type StratejiSchollAppClassesClassDto = {
    id?: string;
    creationTime?: string;
    creatorId?: string | null;
    lastModificationTime?: string | null;
    lastModifierId?: string | null;
    isDeleted?: boolean;
    deleterId?: string | null;
    deletionTime?: string | null;
    tenantId?: string | null;
    name?: string | null;
    level?: string | null;
    isActive?: boolean;
};